Implemented cmake build for IDE.
Only on windows and linux so far, not yet tested on windows.

It also adds a new CMake/PolycodeDependencies.cmake that does most of the work of finding and linking polycode's dependencies for you. New cmake projects can just use this, and if something changes about polycode dependencies, we need only update one file.